+++ disable-blogging.php
@@ -3,11 +3,12 @@
  * Plugin Name: Disable Blogging
  * Plugin URI: https://wordpress.org/plugins/disable-blogging/
  * Description: Turn WordPress into a non-blogging, CMS platform by disabling posts, comments, feeds, and other related the blogging features.
- * Version: 2.0.3
+ * Version: 2.0.4
  * Author: Fact Maven
  * Author URI: https://www.factmaven.com/
  * License: GPLv3
  * Text Domain: dsbl
+ * Domain Path: /languages
  */
 
 # If accessed directly, exit
@@ -17,7 +18,6 @@ if ( ! defined( 'ABSPATH' ) ) exit;
 require_once( ABSPATH . 'wp-admin/includes/plugin.php' );
 # Get plugin's metadata
 $disable_blogging = get_plugin_data( __FILE__ );
-
 # If the plugin version is lower or not defined, remove plugin options
 if ( ( get_option( 'factmaven_dsbl_version' ) < $disable_blogging['Version'] ) || ! get_option( 'factmaven_dsbl_version' ) ) {
     # Remove options with the prefix "factmaven_dsbl_"
@@ -26,16 +26,9 @@ if ( ( get_option( 'factmaven_dsbl_version' ) < $disable_blogging['Version'] ) |
             delete_option( $option );
         }
     }
-    # Delete previous option from v1.3.0
-    delete_option( 'dsbl_remove_profile_fields' );
     # Add options for new plugin version
     update_option( 'factmaven_dsbl_version', $disable_blogging['Version'] );
 }
-
-function dsbl_load_plugin_textdomain() {
-    load_plugin_textdomain( 'dsbl', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
-}
-add_action( 'plugins_loaded', 'dsbl_load_plugin_textdomain' );
 
 # Call the required files
 require_once( plugin_dir_path( __FILE__ ) . 'admin/plugin-meta.php' );